preprocess_song: dynamic spectrogram slicing bug fix where it would never terminate. Also bug fixed continuing tie notes that would be ignored. Also made some better plotting for the verbose
inference: Made a change to music21's own function because it would override pickup measure durations, also bug fixed the bpm token that would be given initially. Went to hell and back in order to figure out why musescore wouldn't display anything. turns out it couldn't handle voices assigned a number of 0. There is no attribute to decide this so we load the xml and change it afterwards (and write a new one) using ET (the library). Chords are also split up to smaller units if the total ambitus is both below octave 4 and above. Voices needed to be implemented because it would otherwise append and make the measures too long. Then we make ties and rests (which are determined from the bar duration attribute). For some reason, the pickup measure duration is ignored during score.write but it is visibly right. so right now it's not a priority to fix